a task can be deleted by its author only (any admin if author's anonymous)

+++ src/AppBundle/Controller/TaskController.php
@@ -3,6 +3,7 @@
 namespace AppBundle\Controller;
 
 use AppBundle\Entity\Task;
+use AppBundle\Entity\User;
 use AppBundle\Form\TaskType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
@@ -87,6 +88,11 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
+        $author = $task->getAuthor();
+        if (null === $author && !$this->isGranted('ROLE_ADMIN') || null !== $author && (!$this->isGranted('IS_AUTHENTICATED_FULLY') || $this->getUser()->getUsername() != $author->getUsername())) {
+            throw $this->createAccessDeniedException();
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
